crear y finalizar consulta fixed
